pyproject.toml
[project]
name = "ivp-section3boafavs"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.2.1",
    "titanoboa>=0.2.8",
    "vyper>=0.4.3",
]
